+++ lib/app_color.dart
@@ -1,40 +1,16 @@
 import 'package:flutter/material.dart';
 
+/// A class that holds commonly used colors for the app.
+///
+/// Provides a centralized place to define theme colors like border, hint text, background, etc.
 class AppColor {
-  static const Color primaryColor = Color(0xFF3A2769);
-  static const Color secondaryColor = Color(0xFF19A1DB);
-  static const Color background = Color(0xffF4F7FF);
+  /// Default border color for text fields and containers.
   static const Color boderColor = Color(0xffECEAF0);
-  static const Color whitecolor =Colors.white;
 
-  // You can define more colors here as needed
-  static const Color textColor = Color(0xFF333333);
+  /// Default hint text color for input fields.
   static const Color hintColor = Color(0xff5F5F5F);
-  static const Color appBlackColor = Color(0xff000000);
-  static const Color cardColor = Color(0xffF4F7FF);
-  static const Color progressBarColor =Color(0xffD9D9D9);
-  static const Color containerColorGredent1 = Color(0xff3A2769);
-  static const Color containerColorGredent2 = Color(0xff6046A3);
-  static const Color containerColorGredent3 = Color(0xff3A2769);
-  static const Color dashboardTextColor = Color(0xff9D93B4);
-  static const Color settlementContianerColor = Color(0xff19A1DB);
-  static const Color searchBarColor = Color(0xffEBECF7);
-  static const Color timeColor = Color(0xff8D8D95);
-  static const Color dividerColor = Color(0xffEBECF7);
 
-
-  static const Color buttonColor = Color(0xFF659287); // The color for buttons
-  static const Color errorColor = Color(0xFFB00020);
-  static const Color theme = Color(0xFfFFFFFF);
-  static const Color bg = Color(0xfffafaef);
-  static const Color backgroundColor = Color(0xff445a6c);
-  static const radioButtonBorderColor = Color(0xff79747E);
-  static const timingContainerColor = Color(0xffFEF1E8);
-  static const containerColor = Color(0xffFFFFFF1A);
-
-
-
-  // rounded container color in transcation Screen
-  static const Color roundCotainerTranscationColor = Color(0xff2B5EC7);
-  static const Color receivedContianerColor = Color(0xff99A4C2);
+// You can define more colors here as needed, e.g.,
+// static const Color primaryColor = Color(0xff123456);
+// static const Color background = Color(0xffffffff);
 }